\name{www.console}
\alias{www.console}
\title{Function to send an expression to remote R engines Consoles}
\description{
  This function sends an expression to remote R engines Consoles.
}
\usage{
  www.console(rlinks,expression)
}
\details{
  Details are given in the package vignette.
}
\value{
  The expression is sent to the consoles of the remote R engines.
}
\author{Karim Chine}
\examples{
  
}
\keyword{console}