% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESalldata.R
\name{es.alldata}
\alias{es.alldata}
\title{Get data limit by size}
\usage{
es.alldata(credential, size)
}
\description{
@param credential the url create by function es.credential
 @param size limit rowsa extracted
}
\details{
@return list with status about connection response
 @export
 @examples
 es.alldata(
 credential = credential,
 size = 100)
}
