% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESgeobuffer.R
\name{es.buffer}
\alias{es.buffer}
\title{Spatial query by a lat/long with buffer defined}
\usage{
es.buffer(credential, variable, size, buffer, longitude, latitude)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{variable}{a vector of variable}

\item{size}{limit rows to extract}

\item{longitude}{longitude}

\item{latitude}{latitude}

\item{bufffer}{radius area (in km)}
}
\value{
data frame
}
\description{
Spatial query by a lat/long with buffer defined
}
\examples{
es.buffer(
 my_credential,
c("var1", "var2"),
100
"1km",
longitude,
latitude
 )

}
