% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESgeobufferring.R
\name{es.buffer.ring}
\alias{es.buffer.ring}
\title{Spatial query by a lat/long between a buffer and a ring}
\usage{
es.buffer.ring(credential, variable, size, latitude, longitude, buffer, ring)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{variable}{a vector of variable}

\item{size}{limit rows to extract}

\item{latitude}{latitude}

\item{longitude}{longitude}

\item{ring}{radius area (in km). ALways plus than buffer}

\item{bufffer}{radius area (in km)}
}
\value{
data frame
}
\description{
Spatial query by a lat/long between a buffer and a ring
}
\examples{
es.buffer(
 my_credential,
c("var1", "var2"),
100
"1km",
"3km",
longitude,
latitude
 )

}
