% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScatcount.R
\name{es.catcount}
\alias{es.catcount}
\title{Agreggating count by a specific variable}
\usage{
es.catcount(credential, variable, size, arrange_count = T)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{variable}{a string variable, must be a factor variable}

\item{size}{limit rows to extract}

\item{arrange_count}{order results by a variable or count. Default T, which mean by count.}
}
\value{
a data frame contains a specific variables
}
\description{
Agreggating count by a specific variable
}
\examples{
 es.catsum(
  my_credential,
  "factor1",
  10,
  T
 )


}
