% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScolumndata.R
\name{es.columndata}
\alias{es.columndata}
\title{Select data by specific variables}
\usage{
es.columndata(credential, size, variable)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{size}{limit rows to extract}

\item{variable}{a vector of specific variables}
}
\value{
a data frame contains a specific variables
}
\description{
Select data by specific variables
}
\examples{
 es.columndata(
  my_credential,
  100,
  c("atributo1", "atributo2" )
 )
}
