% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESconnect.R
\name{es.connection}
\alias{es.connection}
\title{Testing url connection to database with credential}
\usage{
es.connection(credential)
}
\arguments{
\item{credential}{the url create by function es.credential

 @return list with status about connection response
 @export
 @examples
 es.connection(credential)
 )}
}
\description{
Testing url connection to database with credential
}
