% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScredential.R
\name{es.credential}
\alias{es.credential}
\title{Function to create url to connect api elasticsearch}
\usage{
es.credential(
  user = user,
  passwd = passwd,
  find_search = find_search,
  database = database,
  table = table
)
}
\arguments{
\item{user}{username}

\item{passwd}{password}

\item{find_search}{find or search in string}

\item{database}{database name}

\item{collection}{table name}
}
\value{
a character that contain a link
}
\description{
Function to create url to connect api elasticsearch
}
\examples{
 es.credential(
   "entreprise",
   "admin123",
   "search",
   "finance",
   "realties"
 )
}
