% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESmultifilter.R
\name{es.filtermult}
\alias{es.filtermult}
\title{Filtering dataset by a multiple variables and answer. The filter works like a "and operator"}
\usage{
es.filtermult(credential, variable, size, es_filter)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{variable}{a vector of variable}

\item{size}{limit rows to extract}

\item{filter}{list of filters}
}
\value{
a data frame with specific filter
}
\description{
Filtering dataset by a multiple variables and answer. The filter works like a "and operator"
}
\examples{
 es.filterun(
  my_credential,
  c("var1", "var2"),
  100,
  list ("var1" = c("answer1", "answer2"),
         "var2" = c("answer1"))

 )
}
