% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESconvertpolygon.R
\name{es.polygon.r}
\alias{es.polygon.r}
\title{Convert a elastic search polygon to sf object}
\usage{
es.polygon.r(data_geo, var_id, merge_id, data_df)
}
\arguments{
\item{data_geo}{geodataset extracted by elastic search database. Data geo must be a polygon or a multipolygon.}

\item{var_id}{unique vector of geodataset}

\item{merge_id}{name unique vector of geodataset}

\item{data_df}{same geodataset in data_geo, but without geographical atributtes}
}
\value{
a sf object
}
\description{
Convert a elastic search polygon to sf object
}
\examples{
 es.polygon.r(
  city,
  city$id,
  "id",
  city_no_geo
 )
}
