% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESsummarise.R
\name{es.summary}
\alias{es.summary}
\title{Agreggating descritive statics by a specific variable}
\usage{
es.summary(credential, variable)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{variable}{a string variable, must be a numeric variable}
}
\value{
a data frame with statics measures, mean, min and percentile
}
\description{
Agreggating descritive statics by a specific variable
}
\examples{
 es.summary(
  my_credential,
  "numeric_variable"
 )

}
