% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESvariable.R
\name{es.variable}
\alias{es.variable}
\title{Avaliable variables in data collection}
\usage{
es.variable(user, passwd, find_search, database, table)
}
\arguments{
\item{user}{username}

\item{passwd}{password}

\item{find_search}{find or search in string}

\item{database}{database name}

\item{table}{table name}
}
\value{
a vector of all variables names
}
\description{
Avaliable variables in data collection
}
\examples{
 es.credential(
   "entreprise",
   "admin123",
   "search",
   "finance",
   "realties"
 )

}
