% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EStypevariable.R
\name{es.variable.type}
\alias{es.variable.type}
\title{Types of Avaliable variables in data collection}
\usage{
es.variable.type(user, passwd, find_search, database, table)
}
\arguments{
\item{user}{username}

\item{passwd}{password}

\item{find_search}{find or search in string}

\item{database}{database name}

\item{table}{table name}
}
\value{
a named list. list_variable are the variable who are in list. Univariate variables are single variables. Raw variables are variable that need insert a .raw when used for aggregation functions
}
\description{
Types of Avaliable variables in data collection
}
\examples{
 es.credential(
   "entreprise",
   "admin123",
   "search",
   "finance",
   "realties"
 )

}
