% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESalldata.R
\name{es.alldata}
\alias{es.alldata}
\title{Get data limit by size}
\usage{
es.alldata(credential, size)
}
\arguments{
\item{credential}{the url create by function es.credential}

\item{size}{limit rows to extract}
}
\value{
sample data
}
\description{
Get data limit by size
}
\examples{
 es.alldata(
 credential = credential,
 size = 100)

}
