% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESconnect.R
\name{es.connection}
\alias{es.connection}
\title{Testing url connection to database with credential}
\usage{
es.connection(credential)
}
\arguments{
\item{credential}{the url create by function es.credential}
}
\value{
list with status about connection response
}
\description{
Testing url connection to database with credential
}
\examples{
 es.connection(credential)
 )

}
