% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EScredential.R
\name{es.credential}
\alias{es.credential}
\title{Function to create url to connect api elasticsearch}
\usage{
es.credential(
  user = user,
  passwd = passwd,
  find_search = find_search,
  database = database,
  table = table
)
}
\arguments{
\item{user}{username
 @param passwd password
 @param find_search find or search in string
 @param database database name
 @param collection table name

 @return build link
 @export
 @examples
 es.credential(
   "entreprise",
   "admin123",
   "search",
   "finance",
   "realties"
 )}
}
\description{
Function to create url to connect api elasticsearch
}
